digitalocean/requirements.txt
# This requirements file lists all third-party dependencies for this project.
#
# Run 'pip install -r requirements.txt -t lib/' to install these dependencies
# in `lib/` subdirectory.
#
# Note: The `lib` directory is added to `sys.path` by `appengine_config.py`.
Flask==0.10
requests==2.7.0
PyYAML==3.11
networkx==1.9.1
lxml==3.4.4
nltk==3.0.3
simplejson==3.7.3
Scrapy==1.0.1
selenium==2.46.0
